# Conflicts:
#	src/main/java/ru/skillbox/socialnetwork/api/responses/ErrorErrorDescriptionResponse.java
#	src/main/java/ru/skillbox/socialnetwork/model/entity/Comment.java
#	src/main/java/ru/skillbox/socialnetwork/model/entity/Post.java
#	src/main/resources/application.properties

+++ src/main/java/ru/skillbox/socialnetwork/model/entity/Notification.java
@@ -26,7 +26,7 @@ public class Notification {
 //    private Person person;
 
     @Column(name = "entity_id", nullable = false)
-    private Integer entityId;
+    private Long entityId;
 
     @Column(name = "contact", nullable = false, columnDefinition = "varchar(255)")
     private String contact;
